Projet_7_API/requirements.txt
tensorflow
pandas
fastapi
uvicorn
azure-monitor-opentelemetry